+++ game-logic/server/src/main/java/net/bfsr/server/network/NetworkSystem.java
@@ -211,7 +211,7 @@ public class NetworkSystem {
     }
 
     public Packet decodePacket(ByteBuf buffer) throws IOException {
-        int packetId = buffer.readByte();
+        int packetId = buffer.readShort();
 
         try {
             Packet packet = createPacket(packetId);
